--- packages/client/src/app/components/startGame/startGame.tsx
@@ -1,44 +1,0 @@
-/* eslint-disable no-plusplus */
-import React, { FC, useEffect, useState } from 'react';
-import Progressbar from '../common/progressbar/progressbar';
-import style from './startgame.module.scss';
-
-const StartGame: FC = () => {
-    const [value, setValue] = useState(0);
-
-    useEffect(() => {
-        const interval = window.setInterval(() => {
-            if (value >= 100) {
-                clearInterval(interval);
-                return;
-            }
-            setValue(value + 1);
-        }, 30);
-        return () => {
-            clearInterval(interval);
-        };
-    });
-
-    return (
-        <div className={style.startGameBackground}>
-            <h1 className={style.startGameBackground_title}>Get Ready</h1>
-            <div className={style.startGameBackground_footer}>
-                <Progressbar value={value} text="Loading" />
-                {value >= 100 ? (
-                    <a className={style.startGameBackground_footer_playBtn} href="/game">
-                        START
-                    </a>
-                ) : (
-                    ''
-                )}
-            </div>
-            <img
-                className={style.startGameBackground_battleCruiser}
-                src="/src/assets/images/battle-cruiser.png"
-                alt="battle cruiser"
-            />
-        </div>
-    );
-};
-
-export default StartGame;
