+++ packages/client/src/app/pages/landing/landing.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react';
-import StartGame from '@/app/components/startGame/startGame';
+import StartGame from '@/app/pages/startGame/startGame';
 
 const Landing: FC = () => (
     <div>
