+++ packages/client/src/app/pages/startGame/startgame.module.scss
@@ -1,20 +1,20 @@
 @import '@/styles/variables.scss';
 
 @keyframes moveUpDown {
-
-    0%,
-    100% {
-        top: 43%;
-    }
-
-    50% {
-        top: 42%;
-    }
+    0% {
+        transform: translateY(280px);
+      }
+       50% {
+        transform: translateY(300px);
+      }
+      100% {
+        transform: translateY(280px);
+      }
 }
 
 .startGameBackground {
     width: 100%;
-    height: 1080px;
+    height: calc(100vh - 100px);
     background: url('/src/assets/images/start-game-backgound.jpg');
     display: flex;
     justify-content: flex-start;
@@ -31,13 +31,13 @@
     &_battleCruiser {
         position: absolute;
         width: 662px;
-        top: 43%;
+       
         margin-left: auto;
         margin-right: auto;
         left: 0;
         right: 0;
         text-align: center;
-        animation: moveUpDown 7s linear infinite;
+        animation: moveUpDown 7s  ease-in-out infinite both;
     }
 
     &_footer {
@@ -46,7 +46,7 @@
         justify-content: flex-start;
         align-items: center;
         flex-direction: column;
-
+        clear: both;
         &_playBtn {
             font-size: 18px;
             width: 80%;
@@ -58,6 +58,7 @@
             color: #00DFFC;
             border-radius: 50px;
             margin: 40px;
+            z-index: 999;
         }
 
         &_playBtn:hover {
